requirements.txt
# MetaTrader5 is Windows-only; wheel must match the installed Python version.
MetaTrader5>=5.0.45

# GUI — requires Python 3.14 compatible wheel (6.9+ recommended for 3.14)
PySide6>=6.9.0

# Numerics
numpy>=2.1.0
pandas>=2.2.3

# Machine Learning (xgboost removed - compatibility issues with Python 3.14)
scikit-learn>=1.5.0

# Config / validation
pydantic>=2.7.4
pydantic-settings>=2.3.4
PyYAML>=6.0.1

# HTTP / networking
httpx>=0.27.0
aiohttp>=3.10.0
requests>=2.32.3

# Date / time
pytz>=2024.1
python-dateutil>=2.9.0

# Database / ORM
SQLAlchemy>=2.0.30
alembic>=1.13.1

# Logging helpers
structlog>=24.2.0
orjson>=3.10.12

# Testing
pytest>=8.2.2
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0
pytest-cov>=5.0.0
freezegun>=1.5.1
